backend/pyproject.toml
[project]
name = "mechlib"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.41.5",
    "dotenv>=0.9.9",
    "fastmcp>=2.13.2",
    "langchain>=1.0.8",
    "langchain-core>=1.0.7",
    "langchain-google-genai>=3.1.0",
    "langchain-ollama>=1.0.0",
    "langchain-pinecone>=0.2.13",
    "logger>=1.4",
    "pillow>=12.0.0",
    "pinecone-client>=6.0.0",
    "pyexiftool>=0.5.6",
    "fastapi[standard]>=0.124.0",
    "psycopg[binary]>=3.3.2",
    "langchain-postgres>=0.0.16",
    "langchain-openai>=1.1.6",
    "fastmcp>=2.13.2",
]

[project.optional-dependencies]
#  api = [
#
#  ]
#  mcp = [
#
#  ]
  cli = [
      "questionary>=2.1.1",
      "prompt-toolkit>=3.0.52",
  ]
  dev = [
      "pytest>=9.0.1",
  ]
